refactor: replace cross-barrel re-exports with direct imports - W-23008174
