+++ packages/salesforcedx-lwc-language-server/src/lwcServerNode.ts
@@ -13,6 +13,3 @@ export default class Server extends BaseServer {
     return createConnection();
   }
 }
-
-// Re-export for tests
-export { findDynamicContent } from './baseServer';
